tech: extract the estimated durations into a concern
